changed docker-java version

+++ com.chaos-testing-framework/src/main/java/master/com/chaos_testing_framework/service/RollbackContainerService.java
@@ -22,7 +22,7 @@ public class RollbackContainerService {
 
         try {
 
-            Integer cpuQuota = Integer.parseInt(String.valueOf(microserviceMetadata.getCpuQuota()));
+            Long cpuQuota = Long.parseLong(String.valueOf(microserviceMetadata.getCpuQuota()));
 
             Long memoryLimit = microserviceMetadata.getMemoryLimit();
